// worklenz-frontend/src/app/administrator/reporting/interfaces.ts
export interface IRPTOverviewProject {
  id: string;
  name: string;
  client: string;
  status: {
    name: string;
    color_code: string;
    icon: string;
  };
}
// ...
export interface IRPTLastActivity {
  assigned_user?: null;
  attribute_type?: "name";
  current?: string;
  done_by?: {
    name?: string;
    avatar_url?: string;
    color_code?: string;
  };
  avatar_url?: string;
  color_code?: string;
  name?: string;
  log_text?: string;
  log_type?: string;
  previous?: string;
  last_activity_string?: string;
}

export interface IRPTProject extends IRPTOverviewProject {
  last_activity: IRPTLastActivity;
  color_code: string;
  category_id: string | null;
  category_name: string;
  category_color: string;
  team_id: string;
  team_name: string;
  team_color: string;
  status_id: string;
  status_name: string;
  status_color: string;
  status_icon: string;
  start_date: string;
  end_date: string;
  estimated_time: number;
  actual_time: number;
  days_left: number | null;
  is_overdue: boolean;
  is_today: boolean;
  tasks_stat: {
    total: number;
    todo: number;
    doing: number;
    done: number;
  };
  comment?: string;
  project_health: string;
  health_color: string;
  estimated_time_string?: string;
  actual_time_string?: string;
  project_manager: ITeamMemberViewModel
}